[Service Offloading] Load the webrtc server page directly.
Make ChromeTabbedActivity open the webrtc server page at the starting time.
Handling permissions also moves to ChromeTabbedActivity from
AsyncInitializationActivity.
