+++ chrome/android/chrome_java_sources.gni
@@ -1831,13 +1831,13 @@ chrome_java_sources = [
 chrome_java_sources += [
   "java/src/org/chromium/chrome/browser/AlwaysOnTopService.java",
   "java/src/org/chromium/chrome/browser/firstrun/CastanetsFragment.java",
-  "java/src/org/chromium/chrome/browser/init/LicenseAdapter.java",
-  "java/src/org/chromium/chrome/browser/init/SampleAdminReceiver.java",
+  "java/src/org/chromium/chrome/browser/LicenseAdapter.java",
+  "java/src/org/chromium/chrome/browser/SampleAdminReceiver.java",
 ]
 
 if (enable_service_offloading && enable_service_offloading_knox) {
   chrome_java_sources += [
-    "java/src/org/chromium/chrome/browser/init/KnoxLicenseReceiver.java",
-    "java/src/org/chromium/chrome/browser/init/KnoxLicense.java",
+    "java/src/org/chromium/chrome/browser/KnoxLicenseReceiver.java",
+    "java/src/org/chromium/chrome/browser/KnoxLicense.java",
   ]
 }
