+++ chrome/android/java/src/org/chromium/chrome/browser/KnoxLicense.java
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package org.chromium.chrome.browser.init;
+package org.chromium.chrome.browser;
 
 import android.content.Context;
