+++ chrome/android/java/src/org/chromium/chrome/browser/init/AsyncInitializationActivity.java
@@ -8,9 +8,6 @@ import android.Manifest;
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.ApplicationInfo;
-import android.content.pm.PackageManager;
-import android.content.pm.PackageManager.NameNotFoundException;
 import android.content.res.Configuration;
 import android.net.Uri;
 import android.os.Build;
@@ -22,8 +19,6 @@ import android.provider.Settings;
 import android.provider.Settings.SettingNotFoundException;
 import android.support.annotation.CallSuper;
 import android.support.annotation.Nullable;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
 import android.view.Display;
 import android.view.Menu;
 import android.view.View;
@@ -37,7 +32,6 @@ import org.chromium.base.BaseSwitches;
 import org.chromium.base.CommandLine;
 import org.chromium.base.ContextUtils;
 import org.chromium.base.Log;
-import org.chromium.base.OffloadingUtils;
 import org.chromium.base.StrictModeContext;
 import org.chromium.base.TraceEvent;
 import org.chromium.base.VisibleForTesting;
@@ -67,10 +61,6 @@ import org.chromium.ui.modaldialog.ModalDialogManagerHolder;
 
 import java.lang.reflect.Field;
 
-import android.app.Activity;
-import android.app.admin.DevicePolicyManager;
-import android.content.ComponentName;
-
 /**
  * An activity that talks with application and activity level delegates for async initialization.
  */
@@ -80,10 +70,8 @@ public abstract class AsyncInitializationActivity extends ChromeBaseAppCompatAct
     protected final Handler mHandler;
 
     private int ACTION_MANAGE_OVERLAY_PERMISSION_REQUEST_CODE = 5469;
-    private int ACTION_MANAGE_RECORD_AUDIO_PERMISSION_REQUEST_CODE = 1234;
-
-    private String PERMISSON_DENIED_MSG = "The permission was denied. You need to allow the permission to use this app.";
-    private String PERMISSON_DENIED_DONT_ASK_MSG = "The permission was denied. App info > Storage > Manage storage > Clear all data.";
+    protected final String PERMISSON_DENIED_MSG = "The permission was denied. You need to allow the permission to use this app.";
+    protected final String PERMISSON_DENIED_DONT_ASK_MSG = "The permission was denied. App info > Storage > Manage storage > Clear all data.";
 
     private final NativeInitializationController mNativeInitializationController =
             new NativeInitializationController(this);
@@ -117,11 +105,6 @@ public abstract class AsyncInitializationActivity extends ChromeBaseAppCompatAct
     private boolean mFirstDrawComplete;
 
     private Runnable mOnInflationCompleteCallback;
-
-    private static final int DEVICE_ADMIN_ADD_RESULT_ENABLE = 1;
-    private LicenseAdapter mLicenseAdapter;
-    private DevicePolicyManager mDPM;
-    private ComponentName mDeviceAdmin;
 
     public AsyncInitializationActivity() {
         mHandler = new Handler();
@@ -537,18 +520,9 @@ public abstract class AsyncInitializationActivity extends ChromeBaseAppCompatAct
                 // You have permission
                 startService(new Intent(this, AlwaysOnTopService.class));
             } else {
+                Log.e(TAG, "onActivityResult OVERLAY_PERMISSION");
                 showToast(PERMISSON_DENIED_MSG);
                 finish();
-            }
-        } else if (requestCode == DEVICE_ADMIN_ADD_RESULT_ENABLE) {
-            switch (resultCode) {
-                case Activity.RESULT_OK:
-                    mLicenseAdapter.ActivateLicense(this.getApplicationContext());
-                    break;
-                case Activity.RESULT_CANCELED:
-                    showToast(PERMISSON_DENIED_MSG);
-                    finish();
-                    break;
             }
         }
     }
@@ -692,10 +666,6 @@ public abstract class AsyncInitializationActivity extends ChromeBaseAppCompatAct
             }
         }
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-
-        if (OffloadingUtils.IsServiceOffloading()) {
-            checkGrantResults(permissions, grantResults);
-        }
     }
 
     @CallSuper
@@ -903,62 +873,9 @@ public abstract class AsyncInitializationActivity extends ChromeBaseAppCompatAct
                 startService(new Intent(this, AlwaysOnTopService.class));
             }
         }
-
-        if (OffloadingUtils.IsServiceOffloading()) {
-            // Check and request a RECORD_AUDIO permission for service offloading.
-            if (ContextCompat.checkSelfPermission(
-                      ContextUtils.getApplicationContext(),
-                      Manifest.permission.RECORD_AUDIO) != PackageManager.PERMISSION_GRANTED) {
-                Log.w(TAG, "RECORD_AUDIO permission was not granted. Request permission.");
-                ActivityCompat.requestPermissions(this,
-                        new String[] {Manifest.permission.RECORD_AUDIO},
-                        ACTION_MANAGE_RECORD_AUDIO_PERMISSION_REQUEST_CODE);
-            } else {
-                getDeviceAdmin();
-            }
-        }
     }
 
-    private void checkGrantResults(String[] permissions, int[] grantResults) {
-        boolean check_result = true;
-        for (int result : grantResults) {
-            if (result != PackageManager.PERMISSION_GRANTED) {
-                check_result = false;
-                break;
-            }
-        }
-
-        if (check_result) {
-            getDeviceAdmin();
-        } else {
-            if (ActivityCompat.shouldShowRequestPermissionRationale(this, Manifest.permission.RECORD_AUDIO)) {
-                showToast(PERMISSON_DENIED_MSG);
-            } else {
-                showToast(PERMISSON_DENIED_DONT_ASK_MSG);
-
-                Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS,
-                        Uri.parse("package:" + getPackageName()));
-                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                startActivity(intent);
-            }
-            finish();
-        }
-    }
-
-    private void getDeviceAdmin() {
-        mLicenseAdapter = new LicenseAdapter();
-        mDPM = (DevicePolicyManager) getSystemService(Context.DEVICE_POLICY_SERVICE);
-        mDeviceAdmin =
-                new ComponentName(AsyncInitializationActivity.this, SampleAdminReceiver.class);
-
-        // Ask the user to add a new device administrator to the system
-        Intent intent = new Intent(DevicePolicyManager.ACTION_ADD_DEVICE_ADMIN);
-        intent.putExtra(DevicePolicyManager.EXTRA_DEVICE_ADMIN, mDeviceAdmin);
-        // Start the add device administrator activity
-        startActivityForResult(intent, DEVICE_ADMIN_ADD_RESULT_ENABLE);
-    }
-
-    void showToast(String msg) {
+    protected void showToast(String msg) {
         Toast.makeText(this.getApplicationContext(), msg, Toast.LENGTH_LONG).show();
     }
 }
